requirements.txt
boto3==1.34.162
botocore==1.34.162
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
